feat: [SC-25975] :package: Update NameGuard SDK package and app that uses it
