+++ packages/namegraph-sdk/tsup.config.js
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsup";
 
 export default defineConfig({
-  entry: ["src/index.ts"],
+  entry: ["src/index.ts", "src/utils.ts"],
   format: ["esm"],
   clean: true,
   dts: true,
